Included additional stats for 2005-2022
5v5 Time on Ice per Game Played (5v5 TOI/GP), Shot Attempts Percent/CORSI (SAT%), Hits, Hits per 60 minutes (Hits/60), Blocked Shots (BkS), Blocked shots per 60 minutes (BkS/60), Giveaways (GvA), Giveaways per 60 minutes (GvA/60), Takeaways (TkA), Takeaways per 60 minutes (TkA/60), Empty Net Goals (ENG), and Missed Shots (MsS) 

The 2009-2010 season was the first year that included the CORSI (SAT%) metric. I'm not sure if we should guess it for years prior, or ignore it for everybody, or what.
In that vein, I don't have these other stats pre-2005, so it might not even be worth looking at those years.